Refactor tests: move fixtures, be more explicit and prevent intermitent failures
