+++ test/conftest.py
@@ -1,7 +1,12 @@
+from collections import namedtuple
+from unittest.mock import Mock
+
 import pytest
+from eventlet import sleep
 from redis import StrictRedis
 
-from test import REDIS_OPTIONS, URI_CONFIG_KEY
+from nameko_rediskn import rediskn
+from test import REDIS_OPTIONS, TIME_SLEEP, URI_CONFIG_KEY
 
 
 @pytest.fixture
@@ -26,3 +31,41 @@ def redis(redis_config):
     client.flushall()
     yield client
     client.flushall()
+
+
+@pytest.fixture
+def redis_db_1(redis_config):
+    # url argument takes precedence over db in the url
+    redis_uri = '{}?db=1'.format(
+        redis_config['REDIS_URIS'][URI_CONFIG_KEY]
+    )
+    client = StrictRedis.from_url(redis_uri, db=1, **REDIS_OPTIONS)
+    client.flushall()
+    yield client
+    client.flushall()
+
+
+@pytest.fixture
+def tracker():
+    yield Mock()
+
+
+@pytest.fixture
+def create_service(container_factory, config, tracker):
+    def create(config=config, **kwargs):
+        class DummyService:
+
+            name = 'dummy_service'
+
+            @rediskn.subscribe(**kwargs)
+            def handler(self, message):
+                tracker(message)
+
+        ServiceMeta = namedtuple('ServiceMeta', ['container'])
+        container = container_factory(DummyService, config)
+
+        container.start()
+        sleep(TIME_SLEEP)
+        return ServiceMeta(container)
+
+    return create
